requirements.txt
django
pytest
pytest-django
twisted
RPi.GPIO
pytz==2017.2
huey
redis
requests
httmock